Reorganize and tidy
